server/requirements.txt
Flask
Flask-Cors
requests
pusher
python-dotenv
